backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
pydantic>=2.10
pydantic-settings>=2.7
python-dotenv>=1.1
sqlalchemy>=2.0
asyncpg>=0.30
pytest>=8.3
httpx>=0.28
PyJWT[crypto]>=2.10
